Crate identity

IOTA Identity

IOTA Identity is a Rust implementation of decentralized identity, also known as Self Sovereign Identity (SSI), through the W3C Decentralized Identifiers (DID) and Verifiable Credentials standards alongside supporting methods, utilizing the IOTA Distributed Ledger.

Example

use identity::crypto::KeyPair;
use identity::iota::Client;
use identity::iota::Document;
use identity::iota::Network;
use identity::iota::Result;
use identity::iota::TangleRef;

#[smol_potat::main]
async fn main() -> Result<()> {
  // Create a client to interact with the IOTA Tangle.
  let client: Client = Client::new()?;

  // Create a DID Document (an identity).
  let keypair: KeyPair = KeyPair::new_ed25519()?;
  let mut document: Document = Document::from_keypair(&keypair)?;

  // Sign the DID Document with the default authentication key.
  document.sign(keypair.secret())?;

  // Use the client to publish the DID Document to the IOTA Tangle.
  document.publish(&client).await?;

  // Print the DID Document transaction link.
  let network: Network = document.id().into();
  let explore: String = format!("{}/transaction/{}", network.explorer_url(), document.message_id());

  println!("DID Document Transaction > {}", explore);

  Ok(())
}

Output: Example DID Document in the Tangle Explorer.

Documentation & Community Resources

identity.rs: Rust source code of this library on GitHub.

Identity Documentation Pages: Supplementing documentation with simple examples on library usage to get you started.

More Examples: Practical examples to get started with the library.

IOTA Identity Experience Team Website: Website of aforementioned team.

Structure (Temporary)

Resources

Docs Link (Website & User Guide)

X-Team

Simple Example

Architecture/Overview

Get

Modules

core
Core Traits and Types

credential
Verifiable Credentials

crypto
Cryptographic Utilities

did
Decentralized Identifiers

iota
IOTA Tangle DID Method