Module identity::credential

Verifiable Credentials

Specification

Structs

Credential
A Credential represents a set of claims describing an entity.

CredentialBuilder
A CredentialBuilder is used to create a customized Credential.

Evidence
Information used to increase confidence in the claims of a Credential

Policy
Information used to express obligations, prohibitions, and permissions about a Credential or Presentation.

Presentation
A Presentation represents a bundle of one or more VerifiableCredentials.

PresentationBuilder
A PresentationBuilder is used to create a customized Presentation.

Refresh
Information used to refresh or assert the status of a Credential.

Schema
Information used to validate the structure of a Credential.

Status
Information used to determine the current status of a Credential.

Subject
An entity who is the target of a set of claims.

VerifiableCredential
A VerifiableCredential represents a Credential with an associated digital proof.

VerifiablePresentation
A VerifiablePresentation represents a Presentation with an associated digital proof.

Enums

Error
This type represents all possible errors that can occur in the library.

Issuer
An identifier representing the issuer of a Credential.

Type Definitions

Result
Alias for a Result with the error type Error.