Struct identity::crypto::KeyCollection

pub struct KeyCollection {
    type_: KeyType,
    public: Box<[PublicKey], Global>,
    secret: Box<[SecretKey], Global>,
}

A collection of cryptographic keys.

Fields
type_: KeyTypepublic: Box<[PublicKey], Global>secret: Box<[SecretKey], Global>
Implementations

impl KeyCollection

pub fn from_iterator<I>(type_: KeyType, iter: I) -> Result<KeyCollection, Error> where
I: IntoIterator<Item = (PublicKey, SecretKey)>,

Creates a new KeyCollection from an iterator of PublicKey/SecretKey pairs.

pub fn new_ed25519(count: usize) -> Result<KeyCollection, Error>

Creates a new KeyCollection with Ed25519 keys.

pub fn new(type_: KeyType, count: usize) -> Result<KeyCollection, Error>

Creates a new KeyCollection with the given key type.

pub const fn type_(&self) -> KeyType

Returns the type of the KeyCollection object.

pub fn len(&self) -> usize

Returns the number of keys in the collection.

pub fn is_empty(&self) -> bool

Returns true if the collection contains no keys.

pub fn public(&self, index: usize) -> Option<&PublicKey>

Returns a reference to the public key at the specified index.

pub fn public_ref(&self, index: usize) -> Option<KeyRef<'_>>

Returns a KeyRef object referencing the public key at the specified index.

pub fn secret(&self, index: usize) -> Option<&SecretKey>

Returns a reference to the secret key at the specified index.

pub fn secret_ref(&self, index: usize) -> Option<KeyRef<'_>>

Returns a KeyRef object referencing the secret key at the specified index.

pub fn keypair(&self, index: usize) -> Option<KeyPair>

Returns a KeyPair object for the keys at the specified index.

pub fn iter(&self) -> impl Iterator<Item = (&PublicKey, &SecretKey)>

Returns an iterator over the key pairs in the collection.

pub fn iter_public(&self) -> Iter<'_, PublicKey>

Returns an iterator over the public keys in the collection.

pub fn iter_secret(&self) -> Iter<'_, SecretKey>

Returns an iterator over the secret keys in the collection.

pub fn merkle_root<D>(&self) -> Hash<D> where
D: DigestExt,

Returns the Merkle root hash of the public keys in the collection.

pub fn merkle_proof<D>(&self, index: usize) -> Option<Proof<D>> where
D: DigestExt,

Returns a proof-of-inclusion for the public key at the specified index.

pub fn merkle_key_signer<D>(
&'key self,
index: usize
) -> Option<Signer<'static, 'key, Box<dyn SignatureSign + 'static, Global>, D>> where
D: MerkleDigest,

Returns a Merkle Key Signer for the secret key at the specified index.

pub fn merkle_key_verifier<D>(
&self
) -> Verifier<'static, Box<dyn __Target + 'static, Global>, D> where
D: MerkleDigest,

Returns a Merkle Key Verifier for the Merkle root of the key collection.

pub fn encode_key<D>(&self) -> Vec<u8> where
D: MerkleDigest,

Creates a DID Document public key value for the Merkle root of the key collection.