Trait identity::crypto::SignatureName

pub trait SignatureName {
    fn name(&self) -> String;
}

A trait for signature suites identified by a particular name.

Required methods

fn name(&self) -> String

Returns a unique identifier for the signatures created by this suite.

Loading content...