Struct identity::did::Method

pub struct Method<T = ()> {
    id: DID,
    controller: DID,
    key_type: MethodType,
    key_data: MethodData,
    properties: T,
}

A DID Document Verification Method

Fields
id: DIDcontroller: DIDkey_type: MethodTypekey_data: MethodDataproperties: T
Implementations

impl<T> Method<T>

pub fn builder(properties: T) -> MethodBuilder<T>

Creates a MethodBuilder to configure a new Method.

This is the same as MethodBuilder::new().

pub fn from_builder(builder: MethodBuilder<T>) -> Result<Method<T>, Error>

Returns a new Method based on the MethodBuilder configuration.

pub fn id(&self) -> &DID

Returns a reference to the verification Method id.

pub fn id_mut(&mut self) -> &mut DID

Returns a mutable reference to the verification Method id.

pub fn controller(&self) -> &DID

Returns a reference to the verification Method controller.

pub fn controller_mut(&mut self) -> &mut DID

Returns a mutable reference to the verification Method controller.

pub fn key_type(&self) -> MethodType

Returns a reference to the verification Method type.

pub fn key_type_mut(&mut self) -> &mut MethodType

Returns a mutable reference to the verification Method type.

pub fn key_data(&self) -> &MethodData

Returns a reference to the verification Method data.

pub fn key_data_mut(&mut self) -> &mut MethodData

Returns a mutable reference to the verification Method data.

pub fn properties(&self) -> &T

Returns a reference to the custom verification Method properties.

pub fn properties_mut(&mut self) -> &mut T

Returns a mutable reference to the custom verification Method properties.

pub fn try_into_fragment(&self) -> Result<String, Error>