Struct identity::iota::IotaDID

#[repr(transparent)]pub struct IotaDID(DID);

Implementations

impl IotaDID

pub const METHOD: &'static str

The DID method name.

pub const DEFAULT_NETWORK: &'static str

The default Tangle network.

pub fn generate_ed25519<'b, 'c, T, U>(
network: T,
shard: U
) -> Result<(IotaDID, KeyPair), Error> where
T: Into<Option<&'b str>>,
U: Into<Option<&'c str>>,

Generates an IotaDID and KeyPair suitable for ed25519 signatures.

pub fn try_from_borrowed(did: &DID) -> Result<&IotaDID, Error>

Converts a borrowed DID to an IotaDID.

Errors

Returns Err if the input is not a valid IotaDID.

pub fn try_from_owned(did: DID) -> Result<IotaDID, Error>

Converts an owned DID to an IotaDID.

Errors

Returns Err if the input is not a valid IotaDID.

pub unsafe fn new_unchecked_ref(did: &DID) -> &IotaDID

Converts a DID reference to an IotaDID reference without performing validation checks.

Safety

This must be guaranteed safe by the caller.

pub fn parse(input: impl AsRef<str>) -> Result<IotaDID, Error>

Parses an IotaDID from the given input.

Errors

Returns Err if the input is not a valid IotaDID.

pub fn new(public: &[u8]) -> Result<IotaDID, Error>

Creates a new IotaDID with a tag derived from the given public key.

Errors

Returns Err if the input does not form a valid IotaDID.

pub fn with_network<'b, T>(public: &[u8], network: T) -> Result<IotaDID, Error> where
T: Into<Option<&'b str>>,

Creates a new IotaDID for the given public key and network.

Errors

Returns Err if the input does not form a valid IotaDID.

pub fn with_network_and_shard<'b, 'c, T, U>(
public: &[u8],
network: T,
shard: U
) -> Result<IotaDID, Error> where
T: Into<Option<&'b str>>,
U: Into<Option<&'c str>>,

Creates a new IotaDID for the given public key, network, and shard.

Errors

Returns Err if the input does not form a valid IotaDID.

pub fn join(&self, other: impl AsRef<str>) -> Result<IotaDID, Error>

Creates a new IotaDID by joining self with the relative IotaDID other.

Errors

Returns Err if any base or relative DID segments are invalid.

pub fn check_method(did: &DID) -> Result<(), Error>

Checks if the given DID has a valid IotaDID method.

Errors

Returns Err if the input is not a valid IotaDID.

pub fn check_method_id(did: &DID) -> Result<(), Error>

Checks if the given DID has a valid IotaDID method_id.

Errors

Returns Err if the input is not a valid IotaDID.

pub fn check_validity(did: &DID) -> Result<(), Error>

Checks if the given DID is a valid IotaDID.

Errors

Returns Err if the input is not a valid IotaDID.

pub fn is_valid(did: &DID) -> bool

Returns a bool indicating if the given DID is a valid IotaDID.

pub fn network(&self) -> &str

Returns the Tangle network of the IotaDID.

pub fn shard(&self) -> Option<&str>

Returns the Tangle network shard of the IotaDID.

pub fn tag(&self) -> &str

Returns the unique Tangle tag of the IotaDID.

pub fn segments(&self) -> Segments<'_>

pub fn address(&self) -> String

Returns the Tangle address of the DID auth chain.